Fix: 2 1 2025

+++ routes/web.php
@@ -1,16 +1,22 @@
 <?php
 
-use App\Http\Controllers\AdminController;
-use App\Http\Controllers\AdminDataUserController;
-use App\Http\Controllers\AdminLanggananController;
+use App\Models\Rekonsiliasi;
+use App\Mail\GantiPasswordEmail;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\LpjController;
+use App\Http\Controllers\BankController;
 use App\Http\Controllers\BuksController;
+use App\Http\Controllers\UndoController;
+use App\Http\Controllers\UnitController;
+use App\Http\Controllers\AdminController;
 use App\Http\Controllers\BdmukController;
 use App\Http\Controllers\DithnController;
 use App\Http\Controllers\ModalController;
+use App\Http\Controllers\TahunController;
 use App\Http\Controllers\HutangController;
 use App\Http\Controllers\NeracaController;
+use App\Http\Controllers\ProkerController;
 use App\Http\Controllers\ArusKasController;
 use App\Http\Controllers\PiutangController;
 use App\Http\Controllers\ProfileController;
@@ -19,20 +25,19 @@ use App\Http\Controllers\LabaRugiController;
 use App\Http\Controllers\PinjamanController;
 use App\Http\Controllers\BagiHasilController;
 use App\Http\Controllers\InvestasiController;
-use App\Http\Controllers\AktivalainController;
-use App\Http\Controllers\CetakLaporanController;
-use App\Http\Controllers\GantiPasswordController;
 use App\Http\Controllers\LanggananController;
+use App\Http\Controllers\AktivalainController;
 use App\Http\Controllers\PenyusutanController;
 use App\Http\Controllers\PersediaanController;
-use App\Http\Controllers\LaporanLabaRugiController;
-use App\Http\Controllers\LaporanPerubahanModalController;
+use App\Http\Controllers\CetakProkerController;
+use App\Http\Controllers\CetakLaporanController;
 use App\Http\Controllers\RekonsiliasiController;
-use App\Http\Controllers\TahunController;
-use App\Http\Controllers\UndoController;
-use App\Http\Controllers\UnitController;
-use App\Mail\GantiPasswordEmail;
-use App\Models\Rekonsiliasi;
+use App\Http\Controllers\AdminDataUserController;
+use App\Http\Controllers\GantiPasswordController;
+use App\Http\Controllers\AdminLanggananController;
+use App\Http\Controllers\LaporanLabaRugiController;
+use App\Http\Controllers\PenambahanModalController;
+use App\Http\Controllers\LaporanPerubahanModalController;
 
 /*
 |--------------------------------------------------------------------------
@@ -46,7 +51,9 @@ use App\Models\Rekonsiliasi;
 */
 
 Route::get('/', [ProfileController::class, 'index'])->middleware('auth', 'langganan', 'bumdes', 'create.user');
+Route::get('/visi/misi/', [ProfileController::class, 'visiMisi'])->middleware('auth', 'langganan', 'bumdes', 'create.user');
 Route::put('/{profil:id}', [ProfileController::class, 'update'])->middleware('auth', 'langganan', 'bumdes');
+Route::put('/visi/misi/{profil:id}', [ProfileController::class, 'update'])->middleware('auth', 'langganan', 'bumdes');
 
 // Modal
 Route::resource('/modal', ModalController::class)->middleware('auth', 'langganan', 'bumdes');
@@ -128,6 +135,7 @@ Route::get('/export-pdf/rekonsiliasi', [RekonsiliasiController::class, 'exportPd
 // Laporan keuangan neraca
 Route::get('/laporan-keuangan/neraca', [NeracaController::class, 'index'])->middleware('auth', 'langganan', 'bumdes');
 Route::get('/export-pdf/neraca', [NeracaController::class, 'exportPdf'])->middleware('auth', 'langganan', 'bumdes');
+Route::post('/laporan-keuangan/neraca/tutup', [NeracaController::class, 'tutup'])->middleware('auth', 'langganan', 'bumdes');
 
 // Laporan laba rugi
 Route::get('/laporan-keuangan/laporan-laba-rugi', [LaporanLabaRugiController::class, 'index'])->middleware('auth', 'langganan', 'bumdes');
@@ -165,25 +173,61 @@ Route::delete('/admin/data-user/{user:id}', [AdminDataUserController::class, 'de
 
 
 // Langganan Bumdes
-Route::get('/admin/langganan/bumdesa', [AdminLanggananController::class, 'index']);
-Route::get('/admin/langganan/bumdesa/create', [AdminLanggananController::class, 'create']);
-Route::post('/admin/langganan/bumdesa/', [AdminLanggananController::class, 'store']);
-Route::get('/admin/langganan/bumdesa/{langganan:id}/edit', [AdminLanggananController::class, 'edit']);
-Route::put('/admin/langganan/bumdesa/{langganan:id}', [AdminLanggananController::class, 'update']);
-Route::delete('/admin/langganan/bumdesa/{langganan:id}', [AdminLanggananController::class, 'destroy']);
+Route::get('/admin/langganan/bumdesa', [AdminLanggananController::class, 'index'])->middleware(
+    'auth',
+    'admin'
+);
+Route::get('/admin/langganan/bumdesa/create', [AdminLanggananController::class, 'create'])->middleware('auth', 'admin');
+Route::post('/admin/langganan/bumdesa/', [AdminLanggananController::class, 'store'])->middleware('auth', 'admin');
+Route::get('/admin/langganan/bumdesa/{langganan:id}/edit', [AdminLanggananController::class, 'edit'])->middleware('auth', 'admin');
+Route::put('/admin/langganan/bumdesa/{langganan:id}', [AdminLanggananController::class, 'update'])->middleware('auth', 'admin');
+Route::delete('/admin/langganan/bumdesa/{langganan:id}', [AdminLanggananController::class, 'destroy'])->middleware('auth', 'admin');
 
 // Langganan Bumdes Bersama
-Route::get('/admin/langganan/bumdes-bersama', [AdminLanggananController::class, 'index']);
-Route::get('/admin/langganan/bumdes-bersama/create', [AdminLanggananController::class, 'create']);
-Route::post('/admin/langganan/bumdes-bersama/', [AdminLanggananController::class, 'store']);
-Route::get('/admin/langganan/bumdes-bersama/{langganan:id}/edit', [AdminLanggananController::class, 'edit']);
-Route::put('/admin/langganan/bumdes-bersama/{langganan:id}', [AdminLanggananController::class, 'update']);
-Route::delete('/admin/langganan/bumdes-bersama/{langganan:id}', [AdminLanggananController::class, 'destroy']);
+Route::get('/admin/langganan/bumdes-bersama', [AdminLanggananController::class, 'index'])->middleware('auth', 'admin');
+Route::get('/admin/langganan/bumdes-bersama/create', [AdminLanggananController::class, 'create'])->middleware('auth', 'admin');
+Route::post('/admin/langganan/bumdes-bersama/', [AdminLanggananController::class, 'store'])->middleware('auth', 'admin');
+Route::get('/admin/langganan/bumdes-bersama/{langganan:id}/edit', [AdminLanggananController::class, 'edit'])->middleware('auth', 'admin');
+Route::put('/admin/langganan/bumdes-bersama/{langganan:id}', [AdminLanggananController::class, 'update'])->middleware('auth', 'admin');
+Route::delete('/admin/langganan/bumdes-bersama/{langganan:id}', [AdminLanggananController::class, 'destroy'])->middleware('auth', 'admin');
 
 // Email
 Route::post('/kirim-email', [GantiPasswordController::class, 'kirimEmail'])->name('kirim-email');
 Route::get('/ganti-password', [GantiPasswordController::class, 'index'])->name('kirim-email')->middleware('guest');
 Route::get('/kontak/admin', [GantiPasswordController::class, 'kontak'])->name('kontak')->middleware('auth', 'bumdes');
+
+// program kerja
+Route::get('/proker', [ProkerController::class, 'proker'])->middleware('auth', 'langganan', 'bumdes');
+Route::get('/proker/kualititif', [ProkerController::class, 'kualititif'])->middleware('auth', 'langganan', 'bumdes');
+Route::put('/proker/kualititif/{proker:id}', [ProkerController::class, 'kualititifUpdate'])->middleware('auth', 'langganan', 'bumdes');
+Route::get('/proker/strategi', [ProkerController::class, 'strategi'])->middleware('auth', 'langganan', 'bumdes');
+Route::put('/proker/strategi/{proker:id}', [ProkerController::class, 'strategiUpdate'])->middleware('auth', 'langganan', 'bumdes');
+Route::get('/proker/sasaran', [ProkerController::class, 'sasaran'])->middleware('auth', 'langganan', 'bumdes');
+Route::put('/proker/sasaran/{proker:id}', [ProkerController::class, 'sasaranUpdate'])->middleware('auth', 'langganan', 'bumdes');
+Route::get('/proker/rencana/kegiatan', [ProkerController::class, 'rencanaKegiatan'])->middleware('auth', 'langganan', 'bumdes');
+Route::post('/proker/rencana/kegiatan/store', [ProkerController::class, 'kegiatanStore'])->middleware('auth', 'langganan', 'bumdes');
+Route::delete('/proker/rencana/kegiatan/{program:id}', [ProkerController::class, 'kegiatanDestroy'])->middleware('auth', 'langganan', 'bumdes');
+Route::post('/proker/rencana/kerjasama/store', [ProkerController::class, 'kerjasamaStore'])->middleware('auth', 'langganan', 'bumdes');
+Route::delete('/proker/rencana/kerjasama/{kerjasama:id}', [ProkerController::class, 'kerjasamaDestroy'])->middleware('auth', 'langganan', 'bumdes');
+Route::get('/proker/penambahan/modal', [PenambahanModalController::class, 'penambahanModal'])->middleware('auth', 'langganan', 'bumdes');
+Route::post('/proker/alokasi/store', [PenambahanModalController::class, 'alokasiStore'])->middleware('auth', 'langganan', 'bumdes');
+Route::post('/proker/resiko/store', [PenambahanModalController::class, 'resikoStore'])->middleware('auth', 'langganan', 'bumdes');
+Route::post('/update-status', [PenambahanModalController::class, 'updateStatus'])->name('update.status')->middleware('auth', 'langganan', 'bumdes');
+Route::put('/proker/penambahan/modal/{proker:id}', [PenambahanModalController::class, 'penambahanModalUpdate'])->middleware('auth', 'langganan', 'bumdes');
+
+Route::get('/cetak/proker', [CetakProkerController::class, 'exportPdf'])->middleware('auth', 'langganan', 'bumdes')->middleware('auth', 'langganan', 'bumdes');
+
+// LPJ
+Route::get('/lpj', [LpjController::class, 'index'])->middleware('auth', 'langganan', 'bumdes');
+Route::put('/lpj/{lpj:id}', [LpjController::class, 'update'])->middleware('auth', 'langganan', 'bumdes');
+Route::get('/cetak/lpj', [LpjController::class, 'exportPdf'])->middleware('auth', 'langganan', 'bumdes');
+
+// Bank
+Route::resource('/aset/bank', BankController::class)->middleware('auth', 'langganan', 'bumdes');
+Route::post('/aset/bank/update', [BankController::class, 'updateJumlah'])->name('Bank.update')->middleware('auth', 'langganan', 'bumdes');
+Route::get('/aset/export-pdf/bank', [BankController::class, 'exportPdf'])->middleware('auth', 'langganan', 'bumdes');
+Route::put('/aset/rekonsiliasi/bayar/{rekonsiliasi:id}', [BankController::class, 'bayar'])->middleware('auth', 'langganan', 'bumdes');
+
 
 
 // Undo
